requirements.txt
streamlit>=1.35.0
msal>=1.28.0
requests>=2.31.0
reportlab>=4.2.0
python-dotenv>=1.0.0
